add index.ts files, separate out types
