+++ cli.ts
@@ -1,6 +1,6 @@
 import { Command } from "commander";
 
-import { Git } from "./src/models/Git";
+import { Git } from "./src/models";
 import dataSource from "./src/utils/DB";
 
 const pkg: {version: string} = require("../package.json");  // eslint-disable-line
