+++ src/connectors/BaseConnector.ts
@@ -3,11 +3,11 @@ import { existsSync, unlink, writeFileSync } from "fs";
 import * as path from "path";
 
 import { config, hpcConfigMap } from "../../configs/config";
+import { options, hpcConfig, SSH, callableFunction } from "../definitions";
 import FileUtil from "../helpers/FolderUtil";  // shouldn't this be registerUtil?
 import * as Helper from "../helpers/Helper";
 import BaseMaintainer from "../maintainers/BaseMaintainer";
 import { ConnectorError } from "../utils/errors";
-import { options, hpcConfig, SSH, callableFunction } from "../utils/types";
 
 import connectionPool from "./ConnectionPool";
