+++ src/connectors/ConnectionPool.ts
@@ -1,7 +1,7 @@
 import NodeSSH from "node-ssh";
 
 import { config, hpcConfigMap } from "../../configs/config";
-import { SSH, SSHConfig } from "../utils/types";
+import { SSH, SSHConfig } from "../definitions";
 
 // dictionary recording ssh connections for community accounts (which have public ssh ability)
 const connectionPool: Record<string, { counter: number, ssh: SSH }> = {};
