+++ src/connectors/SingularityConnector.ts
@@ -1,11 +1,11 @@
 import * as path from "path";
 
 import { containerConfigMap, hpcConfigMap, kernelConfigMap } from "../../configs/config";
+import { slurm, executableManifest } from "../definitions";
 import * as Helper from "../helpers/Helper";
-import { slurm, executableManifest } from "../utils/types";
 
 import SlurmConnector from "./SlurmConnector";
-// import { kernelConfig } from ".../utils/types";
+// import { kernelConfig } from ".../definitions";
 
 /**
  * Specialization of SlurmConnector that, in addition to supporting ssh/slurm jobs, connects a given singularity container to the HPC environment.
