+++ src/helpers/XSEDEUtil.ts
@@ -1,8 +1,8 @@
 import axios from "axios";
 
 import { config } from "../../configs/config";
-import { Job } from "../models/Job";
-import { hpcConfig } from "../utils/types";
+import { hpcConfig } from "../definitions";
+import { Job } from "../models";
 
 /**
  * Class for accessing XSEDE commands. May be deprecated (https://www.xsede.org/)?
