+++ src/server/GitRoutes.ts
@@ -1,10 +1,10 @@
 import express from "express";
 
+import { executableManifest } from "../definitions";
 import GitUtil from "../helpers/GitUtil";
 import * as Helper from "../helpers/Helper";
-import { Git } from "../models/Git";
+import { Git } from "../models";
 import dataSource from "../utils/DB";
-import { executableManifest } from "../utils/types";
 
 const gitRouter = express.Router();
