+++ src/server/InfoRoutes.ts
@@ -3,10 +3,10 @@ import express from "express";
 import * as fs from "fs";
 
 import { hpcConfigMap, maintainerConfigMap, containerConfigMap, jupyterGlobusMap } from "../../configs/config";
+import { hpcConfig, maintainerConfig, containerConfig, jupyterGlobusMapConfig, announcementsConfig } from "../definitions";
 import * as Helper from "../helpers/Helper";
-import { Job } from "../models/Job";
+import { Job } from "../models";
 import dataSource from "../utils/DB";
-import { hpcConfig, maintainerConfig, containerConfig, jupyterGlobusMapConfig, announcementsConfig } from "../utils/types";
 
 import { authMiddleWare, statistic } from "./ServerUtil";
