+++ src/utils/DB.ts
@@ -5,13 +5,7 @@ import {
 
 import { config } from "../../configs/config";
 import * as Helper from "../helpers/Helper"; 
-import { Cache } from "../models/Cache";
-import { Event } from "../models/Event";
-import { Folder } from "../models/Folder";
-import { Git } from "../models/Git";
-import { GlobusTransferRefreshToken } from "../models/GlobusTransferRefreshToken";
-import { Job } from "../models/Job";
-import { Log } from "../models/Log";
+import { Cache, Event, Folder, Git, GlobusTransferRefreshToken, Job, Log } from "../models";
 
 const entities = [Cache, Event, Folder, Git, GlobusTransferRefreshToken, Job, Log];
