+++ src/utils/Emitter.ts
@@ -1,7 +1,5 @@
 import { config } from "../../configs/config";
-import { Event } from "../models/Event";
-import { Job } from "../models/Job";
-import { Log } from "../models/Log";
+import { Event, Job, Log } from "../models";
 
 import dataSource from "./DB";
