+++ src/utils/Redis.ts
@@ -1,8 +1,8 @@
 import { RedisClientType, createClient } from "redis";
 
 import { config } from "../../configs/config";
-import { Job } from "../models/Job";
-import { credential } from "../utils/types";
+import { credential } from "../definitions";
+import { Job } from "../models";
 
 import dataSource from "./DB";
