+++ src/utils/Supervisor.ts
@@ -4,10 +4,10 @@ import * as events from "events";
 
 import { config, maintainerConfigMap, hpcConfigMap } from "../../configs/config";
 import connectionPool from "../connectors/ConnectionPool";
+import { SSH, callableFunction } from "../definitions";
 import * as Helper from "../helpers/Helper";
 import BaseMaintainer from "../maintainers/BaseMaintainer";
-import { Job } from "../models/Job";
-import { SSH, callableFunction } from "../utils/types";
+import { Job } from "../models";
 
 import dataSource from "./DB";
 import Emitter from "./Emitter";
