+++ tools/globus-refresh-transfer-token.ts
@@ -1,6 +1,6 @@
 import { config, hpcConfigMap } from "../configs/config";
 import PythonUtil from "../src/helpers/PythonUtil";
-import { GlobusTransferRefreshToken } from "../src/models/GlobusTransferRefreshToken";
+import { GlobusTransferRefreshToken } from "../src/models";
 import dataSource from "../src/utils/DB";
 
 const main = async () => {
